ui/requirements.txt
streamlit
streamlit-authenticator
pyyaml
requests
pandas
plotly
